# packages/server/openapi.yaml
openapi: 3.0.3
info:
  title: Campsights API
  version: 1.0.0
  description: API for Campsights app (campsites, elevation, weather)

servers:
  - url: https://campsights.onrender.com/api/v1
    description: Production server (Render)
  - url: http://localhost:3000/api/v1
    description: Development server (Local)

tags:
  - name: Campsites
    description: Operations related to campsite resources
  - name: Health
    description: Health check endpoint
  - name: Auth
    description: User registration and login

paths:
  /campsites:
    get:
      tags:
        - Campsites
      operationId: listCampsites
      summary: Get all campsites
      responses:
        '200':
          description: List of campsites with elevation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Campsite'
              example:
                - id: camp_001
                  name: "Switzerland Trail"
                  description: "Situated amidst towering pines near Sugarloaf Mountain, this campsite provides a peaceful retreat with easy access to hiking trails and scenic overlooks. Ideal for families and adventurers looking for a relaxing getaway."
                  lat: 39.905347
                  lng: -105.57529
                  requires_4wd: false
                  elevation: 2650
                  last_updated: "2025-06-13T21:50:37.842404Z"
                - id: camp_002
                  name: "Hidden Valley"
                  description: "Secluded spot by the creek."
                  lat: 39.7392
                  lng: -104.9903
                  requires_4wd: false
                  elevation: 1800
                  last_updated: "2025-06-19T12:00:00.000Z"
                - id: camp_003
                  name: "Alpine Ridge"
                  description: "High elevation site with panoramic views."
                  lat: 40.015
                  lng: -105.2705
                  requires_4wd: true
                  elevation: 3000
                  last_updated: "2025-07-01T09:30:00.000Z"
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalError'

    post:
      tags:
        - Campsites
      operationId: createCampsite
      summary: Add a new campsite
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CampsiteInput'
            example:
              name: "Hidden Valley"
              description: "Secluded spot by the creek."
              lat: 39.7392
              lng: -104.9903
              requires_4wd: false
      responses:
        '201':
          description: Created campsite
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Campsite'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalError'

  /campsites/{id}:
    get:
      tags:
        - Campsites
      operationId: getCampsiteById
      summary: Get a campsite by ID (with elevation and weather)
      parameters:
        - $ref: '#/components/parameters/campsiteId'
      responses:
        '200':
          description: Campsite with elevation and weather
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CampsiteWithWeather'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalError'

    put:
      tags:
        - Campsites
      operationId: updateCampsite
      summary: Update a campsite
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/campsiteId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CampsiteInput'
            example:
              name: "Hidden Valley"
              description: "Secluded spot by the creek."
              lat: 39.7392
              lng: -104.9903
              requires_4wd: true
      responses:
        '200':
          description: Updated campsite
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Campsite'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalError'

    delete:
      tags:
        - Campsites
      operationId: deleteCampsite
      summary: Delete a campsite
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/campsiteId'
      responses:
        '204':
          description: Deleted successfully
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalError'

  /auth/register:
    post:
      tags:
        - Auth
      summary: Register a new user
      operationId: registerUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthInput'
            example:
              username: "newuser"
              password: "yourpassword"
      responses:
        '200':
          description: JWT token for the new user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthToken'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          description: Username already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          $ref: '#/components/responses/InternalError'

  /auth/login:
    post:
      tags:
        - Auth
      summary: Log in as a user
      operationId: loginUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthInput'
            example:
              username: "newuser"
              password: "yourpassword"
      responses:
        '200':
          description: JWT token for the user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthToken'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          $ref: '#/components/responses/InternalError'

  /health:
    get:
      tags:
        - Health
      operationId: healthCheck
      summary: Health check
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: ok

components:
  parameters:
    campsiteId:
      name: id
      in: path
      description: Unique identifier for a campsite
      required: true
      schema:
        type: string
      example: camp_001

  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    Campsite:
      type: object
      required:
        - id
        - name
        - lat
        - lng
        - requires_4wd
        - elevation
      properties:
        id:
          type: string
          example: camp_002
        name:
          type: string
          example: Switzerland Trail
        description:
          type: string
          example: Situated amidst towering pines near Sugarloaf Mountain, this campsite provides a peaceful retreat with easy access to hiking trails and scenic overlooks. Ideal for families and adventurers looking for a relaxing getaway.
        lat:
          type: number
          format: float
          example: 39.905347
        lng:
          type: number
          format: float
          example: -105.57529
        requires_4wd:
          type: boolean
          example: false
        elevation:
          type: number
          nullable: true
          example: null
        last_updated:
          type: string
          format: date-time
          example: "2025-06-13T21:50:37.842404Z"

    CampsiteInput:
      type: object
      required:
        - name
        - lat
        - lng
        - requires_4wd
      properties:
        name:
          type: string
          example: "Hidden Valley"
        description:
          type: string
          example: "Secluded spot by the creek."
        lat:
          type: number
          format: float
          example: 39.7392
        lng:
          type: number
          format: float
          example: -104.9903
        requires_4wd:
          type: boolean
          example: false

    CampsiteWithWeather:
      allOf:
        - $ref: '#/components/schemas/Campsite'
        - type: object
          properties:
            weather:
              type: array
              items:
                $ref: '#/components/schemas/WeatherPeriod'

    WeatherPeriod:
      type: object
      required:
        - name
        - startTime
        - endTime
        - temperature
        - temperatureUnit
        - windSpeed
        - windDirection
        - shortForecast
        - detailedForecast
      properties:
        name:
          type: string
          example: "Tonight"
        startTime:
          type: string
          format: date-time
          example: "2025-07-03T18:00:00-06:00"
        endTime:
          type: string
          format: date-time
          example: "2025-07-04T06:00:00-06:00"
        temperature:
          type: number
          example: 60
        temperatureUnit:
          type: string
          example: "F"
        windSpeed:
          type: string
          example: "5 mph"
        windDirection:
          type: string
          example: "NW"
        shortForecast:
          type: string
          example: "Clear"
        detailedForecast:
          type: string
          example: "Clear skies tonight with a light breeze."

    AuthInput:
      type: object
      required:
        - username
        - password
      properties:
        username:
          type: string
          example: "newuser"
        password:
          type: string
          example: "yourpassword"

    AuthToken:
      type: object
      required:
        - token
      properties:
        token:
          type: string
          example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."

    Error:
      type: object
      required:
        - error
      properties:
        error:
          type: string

  responses:
    BadRequest:
      description: Bad Request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Not Found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    TooManyRequests:
      description: Too Many Requests
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    InternalError:
      description: Internal Server Error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'